refactor: separate 'pacts for verification' code into separate class
